requirements.txt
# ========================================
# Proxmox 서버 자동 생성 시스템 의존성
# ========================================

# Flask 웹 프레임워크
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Jinja2>=3.1.0
Werkzeug>=2.3.0

# 환경 변수 관리
python-dotenv>=1.0.0

# HTTP 요청
requests>=2.31.0
urllib3>=2.0.0

# 데이터베이스
sqlite3  # Python 내장 모듈

# 보안
cryptography>=41.0.0

# YAML 처리
PyYAML>=6.0

# Vault 통합 (선택사항)
hvac>=1.0.0

# 개발 도구
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# 로깅
colorlog>=6.7.0

# 유틸리티
click>=8.1.0

# 타입 힌트
typing-extensions>=4.0.0